refactor(api): move organization member identity repository to /src/team
