+++ api/lib/domain/usecases/index.js
@@ -64,7 +64,6 @@ import * as membershipRepository from '../../../src/team/infrastructure/reposito
 import { organizationInvitationRepository } from '../../../src/team/infrastructure/repositories/organization-invitation.repository.js';
 import { repositories } from '../../infrastructure/repositories/index.js';
 import { certificationCompletedJobRepository } from '../../infrastructure/repositories/jobs/certification-completed-job-repository.js';
-import * as organizationMemberIdentityRepository from '../../infrastructure/repositories/organization-member-identity-repository.js';
 import * as targetProfileShareRepository from '../../infrastructure/repositories/target-profile-share-repository.js';
 import * as learningContentConversionService from '../services/learning-content/learning-content-conversion-service.js';
 import * as userReconciliationService from '../services/user-reconciliation-service.js';
@@ -133,7 +132,6 @@ const dependencies = {
   organizationInvitationService,
   organizationLearnerIdentityRepository,
   organizationLearnerRepository,
-  organizationMemberIdentityRepository,
   organizationRepository,
   organizationTagRepository,
   organizationValidator,
