+++ api/src/team/infrastructure/repositories/organization-member-identity.repository.js
@@ -1,5 +1,5 @@
-import { knex } from '../../../db/knex-database-connection.js';
-import { OrganizationMemberIdentity } from '../../../src/shared/domain/models/OrganizationMemberIdentity.js';
+import { knex } from '../../../../db/knex-database-connection.js';
+import { OrganizationMemberIdentity } from '../../../shared/domain/models/OrganizationMemberIdentity.js';
 
 const findAllByOrganizationId = async function ({ organizationId }) {
   const sortedMembers = await knex('users')
